+++ src/main/kotlin/org/jetbrains/intellij/IntelliJPluginConstants.kt
@@ -12,6 +12,7 @@ object IntelliJPluginConstants {
     const val BUILD_PLUGIN_TASK_NAME = "buildPlugin"
     const val BUILD_SEARCHABLE_OPTIONS_TASK_NAME = "buildSearchableOptions"
     const val CLASSPATH_INDEX_CLEANUP_TASK_NAME = "classpathIndexCleanup"
+    const val DOWNLOAD_ANDROID_STUDIO_PRODUCT_RELEASES_XML_TASK_NAME = "downloadAndroidStudioProductReleasesXml"
     const val DOWNLOAD_IDE_PRODUCT_RELEASES_XML_TASK_NAME = "downloadIdeaProductReleasesXml"
     const val DOWNLOAD_ROBOT_SERVER_PLUGIN_TASK_NAME = "downloadRobotServerPlugin"
     const val DOWNLOAD_ZIP_SIGNER_TASK_NAME = "downloadZipSigner"
@@ -43,6 +44,7 @@ object IntelliJPluginConstants {
         BUILD_PLUGIN_TASK_NAME,
         BUILD_SEARCHABLE_OPTIONS_TASK_NAME,
         CLASSPATH_INDEX_CLEANUP_TASK_NAME,
+        DOWNLOAD_ANDROID_STUDIO_PRODUCT_RELEASES_XML_TASK_NAME,
         DOWNLOAD_IDE_PRODUCT_RELEASES_XML_TASK_NAME,
         DOWNLOAD_ROBOT_SERVER_PLUGIN_TASK_NAME,
         DOWNLOAD_ZIP_SIGNER_TASK_NAME,